Added Opts 'animOutDelay' and 'animInDelay'
Using the 'fade' transition causes the background to show for a split second between transition from one slide to another. Adding a short delay (animOutDelay) before the current slide transits out allows for the next slide to transit in before the current slide transits out and solves the problem of the background showing for a split second. Added in 'animInDelay' to delay slide transition in - hoping anybody might find a good use for it. but animOutDelay was the one that solved the problem I faced.